unionfind::extra

Trait Extra

pub trait Extra<K, V> {
    type DefaultMappingErr: Error;

    // Required method
    fn default_mapping(
        elems: impl IntoIterator<Item = K>,
    ) -> Result<Self, Self::DefaultMappingErr>
       where Self: Sized;
}

Trait that has to be implemented on types that want to be extra information for each element of a GenericUnionFind.

Required Associated Types

type DefaultMappingErr: Error

Required Methods

fn default_mapping( elems: impl IntoIterator<Item = K>, ) -> Result<Self, Self::DefaultMappingErr>
where Self: Sized,