unionfind

Module generic

Structs

UnionFind

A union find data structure. Note that this implementation clones elements a lot. Generally, you should use the data structure with small, preferably Copyable types, like integers. However, arbitrary Clone+PartialEq types are possible.

Enums

AddError

NewUnionFindError

UnionByRankError

UnionError

UnionOrAddError

UnionStatus

When a union is made, there is a possibility that the two classes were already unioned before. This enum is returned to disambiguate the two cases.