unionfind::mapping

Trait GrowableMapping

pub trait GrowableMapping<K, V>: Mapping<K, V> {
    type AddError: Error + Debug;

    // Required methods
    fn empty() -> Self;
    fn add(&mut self, key: K, value: V) -> Result<(), Self::AddError>;
    fn len(&self) -> usize;

    // Provided methods
    fn set_or_add(&mut self, key: K, value: V) -> Result<bool, Self::AddError> { ... }
    fn is_empty(&self) -> bool { ... }
}

Required Associated Types

type AddError: Error + Debug

Required Methods

fn empty() -> Self

Returns an empty mapping, which can be grown

fn add(&mut self, key: K, value: V) -> Result<(), Self::AddError>

Similar to set, but adds a key to a mapping. Adding may fail due to the implementation being considered full, or the key already being in the set. Note that implementations must return the correct errors to be usable in a union find.

fn len(&self) -> usize

Gets the number of items currently in the mapping.

Provided Methods

fn set_or_add(&mut self, key: K, value: V) -> Result<bool, Self::AddError>

Tries to set the value of the element if the element is already in the set (false is returned). If not, the key is added to the mapping and true is returned.

fn is_empty(&self) -> bool

Checks if the mapping is empty

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.