unionfind::mapping

Trait Mapping

pub trait Mapping<K, V> {
    // Required methods
    fn get(&self, key: &K) -> Option<&V>;
    fn set(&mut self, key: K, value: V);

    // Provided method
    fn contains_key(&self, key: &K) -> bool { ... }
}

A mapping is functionally equivalent to a hashmap. The trait is even implemented for hashmaps. However, in some cases, it’s efficient to use an array instead, since in a union find, keys are often integers, generated in-order.

Required Methods

fn get(&self, key: &K) -> Option<&V>

gets an element from the mapping. Returns none if the key is not found in the mapping.

fn set(&mut self, key: K, value: V)

Set a key to a certain value. The key must be present in the mapping, and callers must check this with get first.

Panics

The implementation may panic when a key cannot be found.

Provided Methods

fn contains_key(&self, key: &K) -> bool

Returns true if the mapping contains a certain element